+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- concord-project-model: `retry` support for tasks;
 - smtp: add support for multiple values in `to`, `cc` and `bcc`
 parameters;
 - concord-server: method to cancel a process including its